feat(coil-extension): wip
