+++ packages/coil-extension/src/types/FrameSpec.ts
@@ -2,3 +2,7 @@ export interface FrameSpec {
   tabId: number
   frameId: number
 }
+
+export function sameFrame(f1: FrameSpec, f2: FrameSpec) {
+  return f1.tabId === f2.tabId && f1.frameId === f2.frameId
+}
